libgcrypt: don't install m4 file into a subdirectory where it can not be found
For example gnome-keyring fails to compile with this as it depends on the libgcrypt.m4
file which is not visible with hiding it in a subdirectory. Furthermore there is no
reason to put the mentioned file in a subdirectory just to hide it as the commit
message for the libgcrypt_1.5.0.bbappend indicates.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>