nodejs: Don't use host python when building
OE-core now stages the python-native binary in a subdirectory that is
not part of PATH by default so that bitbake will always use the host
Python and never python-native. This means that if you have
python-native in DEPENDS it gets built, but not used. What you need to
do instead is inherit from pythonnative, which arranges for
python-native to be built and found in PATH.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>
Change-Id: Ic6ddc38d3964877d1f90db4c135bc4dd319e5fd8
Reviewed-on: http://g2g.palm.com/1103
Reviewed-by: Herb Kuta <herb.kuta@palm.com>
Tested-by: Herb Kuta <herb.kuta@palm.com>