Replace standard boot task with one specific to webos.
Replacing standard boot task with WebOS specific one allows
for customizing the content of webos image, in particular
those responsible for booting, and it is much cleaner implementation.

The problem with starting ls2 services, addressed with this commit,
is the generation of triggers. The presence sysvinit and upstart
interfered with the triggers generation.

Open-webOS-DCO-1.0-Signed-off-by: Iyad Qumei <iyadkq@gmail.com>

Change-Id: I2daf6e0f08bea92e3fb6dd0fbb8605b14f21b975