mm-common: Remove unneeded build dependencies
All this recipe does is stage a tarball and some autotools files, yet
it inherits from gnome.bbclass which adds a build dependency on gconf,
which depends on gtk+, which depends on cairo, virtual/x11 and a raft
of other dependencies that aren't used on the target. Create a
.bbappend for the recipe that eliminates the dependencies, and, since
nothing is compiled, sets PACKAGE_ARCH to "all" by inheriting from
allarch.bbclass .

Thanks to @shr-project for proposing this way of solving the problem.

[OWEBOS-2689] Remove extraneous packages from Open webOS build

Open-webOS-DCO-1.0-Signed-off-by: Herb Kuta <herb.kuta@palm.com>
Change-Id: I1520dce3c8582d276496d2761bf66e100fcc9f07
Reviewed-on: http://g2g.palm.com/941
Tested-by: Herb Kuta <herb.kuta@palm.com>
Reviewed-by: Build Verification
Reviewed-by: Susan Montooth <susan.montooth@palm.com>