webkit-webos: Remove "unusual" use of SRCREV
Although SRCREV is documented as "the revision of the source code used
to build the package", it's expected to used during fetching from a SCM.
This recipe uses it differently: to name the abbreviated commit ID
suffix of the top-level subdirectory found in archive it downloadeds.
Change the recipe to rename the subdirectory to a fixed name, i.e. not
to have the suffix after do_unpack() has completed.

Open-webOS-DCO-1.0-Signed-off-by: Herb Kuta <herb.kuta@palm.com>
Change-Id: I8f209a5e528ff3db97f344250f6b67ffff38e8af
Reviewed-on: http://g2g.palm.com/1071
Tested-by: Herb Kuta <herb.kuta@palm.com>
Reviewed-by: Build Verification
Reviewed-by: Steve Lemke <steve.lemke@palm.com>
Reviewed-by: Susan Montooth <susan.montooth@palm.com>