Add detailed BOM information to buildhistory
:Release Notes:
A full build of webOS will now generate a python "pickle" file that
represents a complete Bill of Materials for all components fetched
in creating the build.

:Detailed Notes:
The information collected by the buildhistory class did not,
for some reason, include the URI from which the code was fetched.
Worse, because it ran during do_package(), it did not record
information for native components or items which were fetched and
built, but not individually packaged (such as static libraries).

This commit Overrides oe-core's buildhistory.bbclass to:
- Save the recipe's SRC_URI in the 'latest' data files during the
  do_package task.
- Create a python pkl file containing a generated BOM for every
  component during do_fetch().

It also adds the scripts/buildhistory-collect-srcuris script to scan
the individual package data files to collect the SRC_URI data
generated during the packaging task. This is done in case any
additional information should be generated after the fetch.

:Testing Performed:
- Performed numerous custom builds, both locally and on Jenkins
- Examined the data files produced to ensure their completeness

:Issues Addressed:
[ES-44] Improve release notes to Include JIRAs etc.
[GF-7662] Add build support for BOM generation

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>
Change-Id: I9e39a9a8f136d60b5c0098abf3c3b4e177416558
Reviewed-on: https://g2g.palm.com/2683
Reviewed-by: Build Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>