app-services: Install script for upstart 1.8
:Release Notes:
Add a couple of lines to install the new etc/init/createLocalAccount.conf
file.  Also change the directory defines used to prevent future trouble.

:Detailed Notes:
The intent of this fix is moving the upstart script - in the
initial implementation of Upstart 1.8, many 'component' upstart jobs
were modified for 1.8 and packaged with starfish-initscripts to ease
the transition from Upstart 0.3.11 to 1.8.  The plan was and is to
move the 'imported' scripts back to their home package as soon as is
reasonable.
This particular fix just adds the lines needed to install the newly
added createLocalAccount.conf in the app-services package in addition
to the existing createLocalAccount upstart 0.3.11 job file.
I also changed the {webos_upstartconfdir} to {sysconfdir}/event.d and
used {sysconfdir}/init for the new file - this is to prevent any
"surprises" if someone changes the webos_upstartconfdir without our
knowledge.

:Testing Performed:
Since the 1.8 file itself is functionally unchanged from that present
in starfish-initscripts, the testing done running Upstart 1.8 + s-i
over the last couple of months still applies.

The testing done specifically for this fix is to locally build the
component with a locally modified copy of meta-webos and inspect
the output directory:
starfish/BUILD-goldfinger/work/all-starfish-linux/app-services/
   3.0.0-1-r6starfish1/image/etc/init and ..../etc/event.d
to confirm that the properly configured upstart scripts end up in their
respective locations.

Mini-BAT performed - results are attached to CCC

:QA Notes:

:Issues Addressed:
[GF-8224] update meta-webos/recipes-webos/app-services/app-services.bb to install upstart 1.8 script
[GF-8223] add Upstart 1.8 script to app-services

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I4aa8fce20c0a6412d60cf99d9b198da2ea6f21be
Reviewed-on: https://g2g.palm.com/2792
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Build Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>