luna-sysmgr-common: Add dependency on binutils-cross
:Release Notes:
g++ won't find ld without this dependency.

:Detailed Notes:
Many recipes require ld from binutils during do_compile, so the right
fix is to include this dependency in BASEDEPENDS, but until this patch:
http://lists.openembedded.org/pipermail/openembedded-core/2013-July/080893.html
is included in our version of oe-core fix it in recipe
(luna-sysmgr-common is the only recipe failing like this in our images).

:Testing Performed:
Builds OK when binutils-cross is built in the same sysroot.

:QA Notes:
No change in image

:Issues Addressed:
[GF-8536] luna-sysmgr-common fails to build separately

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: Ibaf09aa4ac58d3b2c5e8e71df2dd27ad9b2861c2
Reviewed-on: https://g2g.palm.com/2829
Reviewed-by: Build Verification
Reviewed-by: Herb Kuta <herb.kuta@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>