webos-image: adding Upstart 1.8 files only
:Release Notes:
Many Core OS components have had Upstart 1.8 job files added to them
with either no other changes at all or only comments or doc updates
between the previous submission and those in this bug.

:Detailed Notes:
As part of switching to Upstart 1.8, many job files were bundled into
starfish-initscripts with the intention of quickly migrating them
back into the components where they belong.  Many of those components
'belong' to the Core OS group - and this fix is one of three
simultaneous meta updates to handle that migration.

In the components being updated here, the only change between the
previous submission and that included in
conf/distro/include/webos-preferred-versions.inc is the addition
of the Upstart 1.8 job file in the files/launch directory.
The upstart files themselves are just being moved out of
starfish-initscripts - where they've lived since May - and have been
run since then by those of us testing Upstart 1.8

Components updated under meta-webos and the commits included:
* configurator       48  ->  50
  5fd253d add Upstart 1.8 job files
  f60195d Fix API documentation comments
  3d6f831 Add documentation to all service methods
  8eee52b Explicitely link pthread
* filecache          55  ->  56
  49f35d0 Assign Copyright to LG Electronics, Inc.
  ba67e7c filecache: add Upstart 1.8 job file
  d36911f filecache: update upstart script for Upstart 1.8
* luna-sysservice   1.04 ->   5
  e9a4621 add Upstart 1.8 job file
  2197f35 Assign Copyright to LG Electronics, Inc.
  8d61073 Update Cmake script to fix Qt5 build issue
  f1dd5df Removing obsolete conf directory
  4df1644 Use QCoreApplication instead of QApplication
  24c692c Correct includes of lunaservice.h
* pmklogd            10  ->  11
  ad1aae2 Assign Copyright to LG Electronics, Inc.
  05a8182 pmklogd: add Upstart 1.8 job file
* pmlogdaemon       105  -> 106
  82e5390 Assign Copyright to LG Electronics, Inc.
  3671a51 pmlogdaemon: add Upstart 1.8 job file
* powerd             19  ->  20
  38ca7af Assign Copyright to LG Electronics, Inc.
  bf9da0c powerd: add Upstart 1.8 job file
  d98d03a Stop replying on lunaservice.h to include glib.h
* sleepd             14  ->  15
  e6462db Assign Copyright to LG Electronics, Inc.
  b047772 sleepd: add Upstart 1.8 job file
  bf9c90f Clean up comments referring to tellbootie

:Testing Performed:
For this fix, the explicit test was to build an up-to-date local tree
with the new submission numbers for the components involved.
The miniBAT test was run and passed for test cases.
Individually, the submission numbers were tested with local builds
for each component in turn by verifying that the new Upstart job file
and the existing upstart file were both installed into the proper
directory under the 'image/etc' directory under BUILD-goldfinger.
I also verified that the files were in their proper locations in the
package files that were created and in the image when it was expanded
for NFS-mounted testing

:QA Notes:
There are no known QA issues as the files being added with this fix are
not going to be actually used until we turn on Upstart 1.8

:Issues Addressed:
[GF-5620] update coreOS upstart scripts for 1.8
[GF-11494] CCC: configurator=50 filecache=56 luna-sysservice=5
           pmklogd=11 pmlogdaemon=106 powerd=20 sleepd=15
           com.palm.service.palmprofile=58 filenotifyd=66 minicore2=44
           nyx-utils=9 pmlogctld=80 rdxd=128 webos-nettools=3
           luna-sysservice=5

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I172b6c343204ab7de0130037d846dc2b8b1e96a7
Reviewed-on: https://g2g.palm.com/2956
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>