webos_enhanced_submissions: Add sanity check for tags
:Release Notes:
Check if SHA-1 and WEBOS_SUBMISSION set in WEBOS_VERSION correspond
to each other.

:Detailed Notes:
Temporary show only a warning when tag doesn't exist in local checkout,
this will be changed to error later, when all used tags are annotated.
Show error when tag exists, but doesn't match used SHA-1.
This check is executed only when do_unpack is executed, that means that
if someone moves the tag in remote repository, we won't notice it until
do_unpack is re-executed, but that happens on every clean build.

:Testing Performed:
I've set SHA-1 in WEBOS_VERSION to incorrect value and the warning was
shown, I've forced do_unpack to reexecute for recipe with
WEBOS_GIT_REPO_TAG set and all worked fine.

:QA Notes:
No change in image.

:Issues Addressed:
[GF-12496] Introduce sanity checks on WEBOS_VERSION

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I5bc4ad416ce62a700e06f483bac5b2277a5694eb
Reviewed-on: https://g2g.palm.com/3039
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>