webos_enhanced_submissions: Add sanity check for tags
:Release Notes:
When SRCREV as well as submission is set, compare if they correspond
to each other.

:Detailed Notes:
This check is executed only when do_fetch is executed, that means that
if someone moves the tag in remote repository, we won't notice it until
do_fetch is re-executed.

:Testing Performed:
I've set SRCREV to incorrect value and the warning was shown, I've
forced do_fetch to reexecute for recipe with WEBOS_GIT_TAG_CUSTOM and
recipe without SRCREV in WEBOS_VERSION, all worked fine.

:QA Notes:
No change in image.

:Issues Addressed:
[GF-12496] Introduce sanity checks on WEBOS_VERSION

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I5bc4ad416ce62a700e06f483bac5b2277a5694eb