connman=webos3
:Release Notes:
Drop patch which is not needed anymore cause we're starting
wpa-supplicant by dbus auto activation.

:Detailed Notes:
The patch was required as we started wpa-supplicant manually and not
with dbus auto acivation as required by connman. As this is reworked
now we don't have to apply the patch anymore but will continue to
ship as part of our internal connman repository.

:Testing Performed:
Created a complete build and tested it on target device by verifying
that wireless network is still working correctly (scanning, connect
to secure network, disconnect).

:QA Notes:

:Issues Addressed:
[GF-13326] Integrate WiFi direct into the build
[GF-16646] CCC: connman=2

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <simon.busch@lge.com>

Change-Id: I9acb281bf388f43e1373a69b7832b18914905849
Reviewed-on: https://g2g.palm.com/3128
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>