Rename WEBOS_GIT_TAG to WEBOS_GIT_PARAM_TAG
:Release Notes:
Split WEBOS_GIT_TAG into WEBOS_GIT_PARAM_TAG, the value that's passed
via the "tag=" parameter of SRC_URI to the bitbake fetcher, and
WEBOS_GIT_REPO_TAG, which is the name of the tag in the repository from
which the source is meant to be fetched. Right now, they are the same,
and only WEBOS_GIT_PARAM_TAG is set, but in a subsequent commit, they
will be set by webos_enhanced_submissions and will be different.

:Detailed Notes:

:Testing Performed:
Only build tested.

:QA Notes:
No change in image.

:Issues Addressed:
[GF-12492] Convert from setting PREFERRED_VERSION centrally to setting
           WEBOS_VERSION in each recipe
[GF-35136] CCC: Decentralized BOM

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: Ib9aa7afb027206a2ba8b7271983f2d26cafe46b3
Reviewed-on: https://g2g.palm.com/3119
Reviewed-by: Build Verification
Reviewed-by: Iyad Qumei <iyad.qumei@lge.com>
Reviewed-by: Herb Kuta <herb.kuta@lge.com>
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>