webos-image: parted=0
:Release Notes:
Remove parted from the image. It was originally identified as being used
by some end-user component of Open webOS, but this has turned out not to
be the case.

:Detailed Notes:

:Testing Performed:
"make webos-image" succeeded and confirmed that parted was not longer
present in the image.

:QA Notes:

:Issues Addressed:
[GF-4063]  Remove GPLv3 components from webOS end-user images
[GF-39419] CCC: parted=0; move binutils developer image

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: Ibe00fefa70fd1fb374723034f5f6475eba50df97
Reviewed-on: https://g2g.palm.com/3140
Reviewed-by: Build Verification
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>