webos_machine_impl_dep: Set PACKAGE_ARCH to MACHINE_ARCH
:Release Notes:
If there are MACHINE-s with different "machine implementations" but
with the same TUNE_PKGARCH, they will have the same package name unless
PACKAGE_ARCH is set to MACHINE_ARCH.

:Detailed Notes:
Is there a better way to handle this?

:Testing Performed:
Built novacomd with this change, and its package name was changed as
expected.

:QA Notes:

:Build Notes:
The package names of all of the components that are "machine
implemention" dependent will change.

:Issues Addressed:
[GF-33432] Correct the PACKAGE_ARCH of components that are "machine
           implementation" dependent

Open-webOS-DCO-1.0-Signed-off-by: Herb Kuta <herb.kuta@lge.com>
Change-Id: Idfd9a61c86f7b2cce6ef2ec2750ee9a0414447d4
Reviewed-on: https://g2g.palm.com/3208
Reviewed-by: Herb Kuta <herb.kuta@lge.com>
Tested-by: Herb Kuta <herb.kuta@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>