packagegroup-*: Add comment about changed policy on PR bumps
:Release Notes:
PR bumps are no longer required when the runtime dependencies are
changed in packagegroup recipes. Add a reminder of this as a comment.

:Detailed Notes:
The reason for the exception for packagegroup recipes is to reduce the
conflicts when merging changes that add or remove packages from the images.
Having PR change when this happens is not necessary.

:Testing Performed:
Only parsed.

:QA Notes:
No change in image or build, only comment was added.

:Issues Addressed:
[GF-13891] Drop requirement that PR be bumped in packagegroup recipes
[GF-35136] CCC: Decentralized BOM

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I9c7dbc0cfebf25690a0dce82877984f3b85d2d41
Reviewed-on: https://g2g.palm.com/3335
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>