libnih: Use BBCLASSEXTEND and fix sstate signature
:Release Notes:
libnih-native was depending on target dbus. This was causing a failure
in the configuring of libnih-native when building upstart in a clean
tree. It was also causing the sstate signature of libnih-native to be
dependent upon the target device.

:Detailed Notes:

:Testing Performed:
Confirmed that a fresh build of upstart succeeded and checked with
sstate-diff-machines.sh.

:QA Notes:
No change in image.

:Issues Addressed:
[GF-37190] libnih-native has incorrect dependency on non-native dbus

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I906fd0e714ccd1d49257a4c828b4bc1d808504cb
Reviewed-on: https://g2g.palm.com/3360
Reviewed-by: Herb Kuta <herb.kuta@lge.com>
Tested-by: Herb Kuta <herb.kuta@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>