packagegroups-webos-{extended,test}: Move binutils to developer image
:Release Notes:
Binutils is not needed in the end-user image, but it's still useful to
have it available in the developer image (e.g. objdump).

:Detailed Notes:
binutils was originally identified as being used by some end-user
component of Open webOS, but this has turned out not to be the case.

:Testing Performed:
Built end-user image without binutils, booted target and performed basic
testing.

:QA Notes:

:Issues Addressed:
[GF-3170]  Determine if binutils can be removed from webOS images
[GF-4063]  Remove GPLv3 components from webOS end-user images
[GF-39419] CCC: parted=0; move binutils developer image

Open-webOS-DCO-1.0-Signed-off-by: Iyad Qumei <iyad.qumei@lge.com>
Change-Id: I580ddaa2b15d3267fa15bd050551df763861036f
Reviewed-on: https://g2g.palm.com/3486
Reviewed-by: Build Verification
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>