upstart: Fix memory leak by turning off logging
:Release Notes:
A memory leak was found that traced to Upstart's attempts to log job
output to non-existent log file, which caused Upstart to buffer the data
with no way to clear it.

:Detailed Notes:
Upstart's defaults include logging the output of all job files to a
number of files located in /var/log/upstart/<jobname>.log.  Since we do
not create /var/log/upstart, it was apparently failing the create, but
then continuing to try to log the data, and when it couldn't open the
log file, it was buffering the data (often the same block again and
again) to a never-ending increasing sized buffer.  This is ... bad.
We did not know that Upstart was trying to log, and we do not want it
to log, so the quick fix for us is to simply change the default to
not do logging.  As we do not want Upstart logging, there is no added
mechanism to override this setting.
I also moved the upstart 1.8 specific patch files to the upstart-1.8
directory as they are unlikely to work on any other version.

:Testing Performed:
Locally built - overnight memory logging run show zero increase in
usage by init. (Previous tests showed about 100kB per minute
increases)
MiniBAT run on flashed TV using gf build 1472.  Results attached
in comment of CCC GF-41569

:QA Notes:

:Issues Addressed:
[GF-37332] Upstart memory consumption keeps going up when TV is idle
[GF-41569] CCC: upstart=r4

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: Id782d5b0b5e53fbd57ed9bb682c9cf51c8c9d928
Reviewed-on: https://g2g.palm.com/3620
Reviewed-by: Jongho Kim <jongho1208.kim@lge.com>
Tested-by: Jongho Kim <jongho1208.kim@lge.com>