Upstart: Fix memory leak by turning off logging
:Release Notes:
A memory leak was found that traced to Upstart's attempts to log job
output to non-existent log file, which caused Upstart to buffer the data
with no way to clear it.

:Detailed Notes:
The previous fix for this problem somehow managed to not include
the actual recipe file with the inclusion of the new patch.

:Testing Performed:
miniBAT run on verification build 1488 on gf.  Also ren top several
times during the test and saw some variation (depending on what else
was running on the device) but no change in overall memory use:
Output of top -bn1 -p1:
after boot:
  PID %CPU %MEM  RES  VIRT SWAP  SHR CODE DATA COMMAND
    1    0  0.2 2304  3796 1492 1252  216 1232 init
after miniBAT:
    1    0  0.1 1360  3796 2436  744  216 1232 init
do an initctl list
    1    0  0.2 2016  3796 1780 1096  216 1232 init
run 7 apps:
    1    0  0.2 1964  3796 1832 1068  216 1232 init

:QA Notes:

:Issues Addressed:
[GF-37332] Upstart memory consumption keeps going up when TV is idle
[GF-41569] CCC: Upstart=r4

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: Ic8466e319b4a4da61118552636c8ec23816c1093
Reviewed-on: https://g2g.palm.com/3637
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>