webos-image: upstart=1.10
:Release Notes:
Upstart 1.10 adds some bug fixes and restores libupstart which was
removed in Upstart 0.6.0

:Detailed Notes:
Moving to Upstart 1.10 will be invisible to most as the API remains
exactly the same as it is in Upstart 1.8.
The list of installed files is the same as for the current
Upstart 1.8 - with all being located in /sbin.
Version 1.10 re-introduces libupstart which is packed in three new
packages.

:Testing Performed:
In addition to a miniBAT, long-term memory use is the same as
it was in 1.8 with no increase past initial amount.
Power cycles: 1121 power cycles with zero failures
Starting and stopping: testing many jobs passed (stopping all jobs
  is not practical as that will essentially kill the TV)

:QA Notes:

:Issues Addressed:
[GF-43901] Upgrade to Upstart 1.10
[GF-44834] CCC: upstart=v1.10

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I9dc201fd3d9953af33c310ecb7fe4a1094cbde46
Reviewed-on: https://g2g.palm.com/3737
Reviewed-by: Build Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>