webos.inc: Set umask 022 for do_unpack task
:Release Notes:
Different umask on different jenkins slaves was causing a lot of
changes in files-in-image.txt report.

:Detailed Notes:
For details see:
https://bugzilla.yoctoproject.org/show_bug.cgi?id=5590

:Testing Performed:
Only build tested and verified in buildhistory, that 2 builds executed
with different system umask now produce the same permissions in
files-in-image.txt buildhistory report.

:QA Notes:
A lot of files and directories will loose writable bit for groups.
It was like that in some builds, now it will be determinictic and
the same in all builds.

:Issues Addressed:
[GF-14832] Determine if there are other native tools that must be
           changed to be dynamically linked
[GF-52622] Resolve random changes in files-in-image.txt report

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>

Change-Id: Ie902f33189724b07518dfcab823b5ef83601ae42
Reviewed-on: https://g2g.palm.com/4017
Reviewed-by: DCO Verification
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>