webos.inc: Change definition of WEBOS_DISTRO_RELEASE_CODENAME
:Release Notes:
Consider the definition of WEBOS_DISTRO_RELEASE_CODENAME to contain two
fields: WEBOS_DISTRO_PLATFORM_CODENAME and WEBOS_DISTRO_BUILD_CODENAME,
separated by a hyphen.

:Detailed Notes:
WEBOS_DISTRO_PLATFORM_CODENAME names the major set of features of the
platform and WEBOS_DISTRO_BUILD_CODENAME names the series of builds that
expected that there will be several build codenames with the same
platform codename under active development at the same time. They
roughly correspond the the major and minor versions of DISTRO_VERSION,
but allow there to be a name to refer to them ahead of the assignment of
actual version numbers.

The defaults for these new variables have been set so that the value of
WEBOS_DISTRO_RELEASE_CODENAME is unchanged.

:Testing Performed:
Confirmed using:
  $ bitbake -e libpbnjson | grep 'WEBOS_DISTRO_.*_CODENAME='
that the values of these variables are as expected.

:QA Notes:

:Issues Addressed:
[GF-57856] Change definition of WEBOS_DISTRO_RELEASE_CODENAME
[GF-58280] CCC: Add support for branch builds

Open-webOS-DCO-1.0-Signed-off-by: Herb Kuta <herb.kuta@lge.com>

Change-Id: I6785318aa6da92472673dbd04262e0a92d889ca6
Originally-reviewed-on: https://g2g.palm.com/4374
Cherry-picked-from-commit: c9bd462e2f6a9c7c77dc08edf7e2653fc4a512a9
Cherry-picked-from-branch: @acadia
Reviewed-on: https://g2g.palm.com/4395
Reviewed-by: DCO Verification
Reviewed-by: Herb Kuta <herb.kuta@lge.com>
Tested-by: Herb Kuta <herb.kuta@lge.com>
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>