webos_enhanced_submissions: Check that SHA-1 is included in right branch
:Release Notes:
Check that the revision-hash included in WEBOS_VERSION is contained in
the branch specified by the branch parameter in SRC_URI.

:Detailed Notes:
The branch parameter isn't very important in the current bitbake (unless
AUTOREV is begin used), but the latest bitbake from upstream checks that
the revision-hash specified is contained in the specified branch and
fails if it's not. See:
http://lists.openembedded.org/pipermail/bitbake-devel/2013-December/004212.html

:Testing Performed:
Only build tested and verified that error is shown for components
with known issues.

:Build Notes:
All components with known issues need to be fixed before this is merged.

:QA Notes:
No change to image.

:Issues Addressed:
[GF-59659] Allow the branch to be specified in WEBOS_VERSION
[GF-59660] CCC: Allow the branch to be specified in WEBOS_VERSION

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Open-webOS-DCO-1.0-Signed-off-by: Susan Montooth <susan.montooth@lge.com>

Originally-reviewed-on: https://g2g.palm.com/4521
Cherry-picked-from-commit: 65e817c441aa0a419947044942ea5103dfc77b0f
Cherry-picked-from-branch: @acadia

Change-Id: I43d4a167728b048c879377f83c1b5da4b1cda94f