package.bbclass: Import fix for newer tar 1.27
:Release Notes:
package.bbclass in oe-core/dylan already has this fix backported, but
because we have own copy, we need to update it as well.

:Detailed Notes:
do_package was failing with newer tar 1.27 because we were missing this
change:
http://git.openembedded.org/openembedded-core/diff/meta/classes/package.bbclass?h=dylan&id=5e3ec61bacd393d461498d9e090306cc7bd8fa63
It is still compatible with older tar.

tar version 1.27 returns:
tar: --same-order option cannot be used with -c

with the commandlines we have been using. We can remove the -s option (which
is --same-order) to remove the error.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Build Notes:
Invalidates almost all sstate signatures.

:Issues Addressed:
[ES-620] Fix current metadata to build correctly on 14.04 Ubuntu

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: Ia5013a98c41025c0b947909dc3bde38acad4cf6f
Reviewed-on: https://g2g.palm.com/5000
Reviewed-by: DCO Verification
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Steve Lemke <steve.lemke@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>