webos.inc: Invent WEBOS_DISTRO_VERSION_CODENAME
:Release Notes:
Create a new codename WEBOS_DISTRO_VERSION_CODENAME that represents the
features of a particular DISTRO_VERSION so that it can be different from
WEBOS_DISTRO_BUILD_CODENAME, which appears in the image filenames.

:Detailed Notes:
In the cases where sustaining branches are made off a master branch and
where release branches are made off a sustaining branch, the build of
the master and sustaining branches will contain different features at
different points in time. This means that WEBOS_DISTRO_BUILD_CODENAME
can't be used to represent the features of a release. Invent
WEBOS_DISTRO_VERSION_CODENAME to represent the features of a release,
leaving WEBOS_DISTRO_BUILD_CODENAME to be used to form the image
filenames.

:Testing Performed:
Confirmed using:
  $ bitbake -e starfish-image | \
    egrep '^WEBOS_DISTRO_(BUILD|VERSION)_CODENAME='
that the settings are correct and using
  $ bitbake -e starfish-image | grep '^IMAGE_NAME='
that the correct codename is used to form IMAGE_NAME.

:QA Notes:

:Issues Addressed:
[ANG-3324] Wrong codename appears in image filenames
[ANG-3421] CCC: Invent WEBOS_DISTRO_VERSION_CODENAME

Open-webOS-DCO-1.0-Signed-off-by: Herb Kuta <herb.kuta@lge.com>
Open-webOS-DCO-1.0-Signed-off-by: Susan Montooth <susan.montooth@lge.com>

Originally-reviewed-on: https://g2g.palm.com/4985
Cherry-picked-from-commit: edfc16c341a66a0913fa393dab6d8928b6cc08ef
Cherry-picked-from-branch: @afro

Change-Id: I606e1e9f647ad95d34446322a6460698ee29dea4