webos_enhanced_submissions.bbclass: Run git remote update when sanity fails
:Release Notes:
Move sanity check for tags and branches to separate function and when it
fails in a way which can be resolved by git remote update, call it and
execute the sanity check again.

:Detailed Notes:
When some SRCREV is merged to correct branch after the build already
populated the premirror, the fetcher doesn't call git pull to discover
that now the same commit is also in other branch and sanity check can
fail. The same can happen with the tag, when tag is created and pushed
before the branch.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[ES-862] Clarify documentation for creating pre-submission builds
[ES-882] Fix submissions bbclasses to support pre-submission builds
[BHV-9439] CCC: Fix submissions bbclasses to support pre-submission and
           multiple levels of branched submissions

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I692b84703b559de09cd3c3a805ee8c9146ab9a57
Reviewed-on: https://g2g.palm.com/5752
Reviewed-by: DCO Verification
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>