webos_enhanced_submissions: Fix various SRCREV sanity check failures
:Release Notes:
Fix build errors due to sanity check failures associated the SRCREV
extracted from WEBOS_VERSION having values of AUTOINC or INVALID.

:Detailed Notes:
Fix build error due to specifying ${AUTOREV} for SRCREV:

 * ERROR: QA Issue: Revision 'AUTOINC' defined in WEBOS_VERSION for
   recipe 'component' (file '<path to recipe>/*.bb') isn't included in
   branch 'develop'

Fix build error due to specifying a tag not associated with a
particular branch, such as a code review on Gerrit:

 * ERROR: QA Issue: Revision 'INVALID' defined in WEBOS_VERSION for
   recipe 'component' (file '<path to recipe>/component.bb') isn't
   included in branch 'master'

This fix provides an alternative to having to set WEBOS_GIT_BRANCH to
";nobranch=1". (But since we will eventually migrate to the upstream
version of the bitbake fetcher that will require this, folks should get
into the habit of setting it.)

:Testing Performed:
Successfully built components in the scenarios specified in the details
section.

:QA Notes:

:Issues Addressed:
[ES-883] Rewrite documentation and fix webos_enhanced_submissions

Open-webOS-DCO-1.0-Signed-off-by: Iyad Qumei <iyad.qumei@lge.com>

Change-Id: I4b615ef5f80078d0656bc1a4d5273117902a310a
Reviewed-on: https://g2g.palm.com/5718
Reviewed-by: DCO Verification
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>